requirements.txt
requests
fastapi
uvicorn
pydantic
python-dotenv
